split event_handler service
